+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -2,6 +2,10 @@
 
 Release History
 ===============
+unreleased
++++++++++++++++++++
+* correct sshMaster0 port number for kubernetes
+
 2.0.12 (2017-08-11)
 +++++++++++++++++++
 * add preview regions